+++ projects/ngx-mat-toast/src/lib/toast-container/toast-container.component.spec.ts
@@ -38,7 +38,9 @@ describe('ToastContainerComponent', () => {
       dismiss: (id: string): void => {
         dismissedIds.push(id);
       },
+      tapped: (_id: string): void => {},
       position: { horizontal: 'start', vertical: 'top' },
+      fullWidth: false,
     };
 
     await TestBed.configureTestingModule({